Visitor detection and additonal Micro:Bit functionality
Visitor block added to the Quando menu. A remote Micro:Bit's proximity to the connected Micro:Bit can be detected, this is categorised as either "close to the exhbit" or "far away from the exhibt".

Additonal functionality added to the Micro:Bit block. If a connected Micro:Bit is connected/disconnected from the server, and if a remote Mirco:Bit is paired with the connected Micro:Bit.