visual recognition api working ish
api call works, but I cant figure out how to send it through to the client instead of just logging it to the console...

things i need to think about:
- should the visual recognition be sampling the video at intervals or only being called upon an event, like the user pressing a button to confirm. Former would be a lot of overhead, but must be possible, latter will require a lot of UI work....

Next is draco compression, will Quando have to autocompress everything as it's uploaded, on startup of server, or should it be up to the user?

Asset management, this would allow for a total scene loading screen, but I think that a model based one is better. Need to find out if a-entitiy will emita  loaded event, else I'll have to go with management. umm ok think that's all for now